refactor(frontend): clean dead draft UsersPage, rename UsersMan to UsersPage
